readability: remove redundant casts (readability-redundant-casting)
Automated fix via clang-tidy readability-redundant-casting.
41 files — removes casts that convert a value to its own type.

+++ indra/llcorehttp/_httpoprequest.cpp
@@ -864,7 +864,7 @@ int HttpOpRequest::seekCallback(void *userdata, curl_off_t offset, int origin)
         return 2;
     }
 
-    op->mCurlBodyPos = (size_t)newPos;
+    op->mCurlBodyPos = newPos;
 
     return 0;
 }
